+++ src/result/__init__.py
@@ -1,4 +1,4 @@
-from .result import Err, Ok, OkErr, Result, UnwrapError, as_result
+from .result import Err, Ok, OkErr, Result, UnwrapError, as_result, is_err, is_ok
 
 __all__ = [
     "Err",
@@ -7,5 +7,7 @@ __all__ = [
     "Result",
     "UnwrapError",
     "as_result",
+    "is_err",
+    "is_ok",
 ]
 __version__ = "0.8.0"